test: extract menu-bar overflow tests to separate suite
